Revert "driver: dw_mmc: add work routine for sudden change"
This reverts commit 2c47c8325e3597518878e41af81110bbaa0b257a.

This patch was merged when there was no timeout timer for command.(CTO)
So at that time, when host couldn't get respond of command, the system was hung.
In that case, this work routine help the system. But there is CTO
in kernel mainline now and CTO is triggered in that case. Both of CTO and
work routine are triggered. So we don't need work routine anymore.

Change-Id: I91210edb525750b33c8fb584458523c461358d01
Signed-off-by: Sowon Na <sowon.na@samsung.com>
Signed-off-by: Sungwoo Park <swpark@nexell.co.kr>
Reviewed-on: https://review.gerrithub.io/434952