[UPSTREAM] binfmt_elf: compat: avoid unused function warning
When CONFIG_ELF_CORE is disabled, we get a harmless warning in the compat
version of binfmt_elf:

fs/compat_binfmt_elf.c:58:13: error: 'cputime_to_compat_timeval' defined but not
used [-Werror=unused-function]

This was addressed in mainline Linux as part of a larger rework with commit
cd19c364b313 ("fs/binfmt: Convert obsolete cputime type to nsecs").

For 4.9 and earlier, this just shuts up the warning by adding an #ifdef
around the function definition.

Change-Id: If95d80a5fe3d40506ba320465845b159bf246ba6
Signed-off-by: Sungwoo Park <swpark@nexell.co.kr>