dtsi: s5p6818_avn_ref: add logical video nodes
This patch adds a logical clipper and decimator node
interfaced with mipi and compatible with tw9992 sensor.
It's added as a logical device to use the same vip port 0
with clipper_0. so it creates two video device using the same port.
(like video6, video7)
each video device can't compatible with the other.
both video device can't be used together.

Change-Id: Ie68f7b2c1553890b1380f61898766eb2ee2a6150
Signed-off-by: HyejungKwon <cjscld15@nexell.co.kr>
Reviewed-on: https://review.gerrithub.io/c/437792
Reviewed-by: Sungwoo Park <swpark@nexell.co.kr>
Tested-by: Sungwoo Park <swpark@nexell.co.kr>