media: nexell: clipper: clear vip stuck issue
This patch clears a vip stuck issue generated when
two vip modules are being tested to start and stop stream
repeatedly at the same time.
It is related to tasklet for buffer handling and waiting scenario
for stopping.
so it removes tasklet and change the order to set the stop flag.

Signed-off-by: HyejungKwon <cjscld15@nexell.co.kr>
Change-Id: I0b54001d05079a1b11d1a86cb47324eb8976156f
Reviewed-on: https://review.gerrithub.io/c/NexellCorp/linux_kernel-4.4.x/+/461188
Reviewed-by: Sungwoo Park <swpark@nexell.co.kr>
Tested-by: Sungwoo Park <swpark@nexell.co.kr>