media: nexell: decimator: clear vip stuck issue
This patch clears a vip stuck issue generated when
two vip modules are being tested to start and stop stream
repeatedly at the same time.
It is related to tasklet for buffer handling and waiting scenario
for stopping.
so it removes tasklet and change the order to set the stop flag.

Signed-off-by: HyejungKwon <cjscld15@nexell.co.kr>
Change-Id: Ifc914acb5e9ed2f26ca10e92a3f340843b126da1